# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DOC Intelligence API
  version: "0.1.0"
  description: >
    Contrato da fatia vertical (Trilha A). Formalização de docs/spec.md §5 — em caso de
    divergência entre este arquivo e o comportamento real do código, o código manda; abra uma
    ADR registrando a diferença.
servers:
  - url: http://localhost:3000/v1
security:
  - apiKey: []
paths:
  /documents:
    post:
      summary: Recebe um documento novo
      security: [{ apiKey: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF, JPEG ou PNG — tipo real detectado pelo servidor por magic bytes.
                channel:
                  type: string
                  enum: [whatsapp, email, balcao, outro]
                  default: outro
                sourceFilename:
                  type: string
                  description: Nome enviado pelo cliente. Nunca usado para decidir tipo/conteúdo.
      responses:
        "201":
          description: Documento novo, criado com status "received".
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
        "200":
          description: Hash já existia — devolve o documento existente, com duplicate=true.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
        "400":
          description: >
            Erro do cliente, diferenciado por `error.code`: `missing_file` (campo file ausente),
            `file_too_large` (acima de MAX_UPLOAD_BYTES) ou `unsupported_file_type` (magic bytes
            não batem com PDF/JPEG/PNG).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code: { type: string, enum: [missing_file, file_too_large, unsupported_file_type] }
                      message: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
    get:
      summary: Lista documentos
      security: [{ apiKey: [] }]
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/Status" } }
        - { name: docType, in: query, schema: { $ref: "#/components/schemas/DocType" } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        "200":
          description: Página de documentos.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Document" } }
                  page: { type: integer }
                  pageSize: { type: integer }
                  total: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /documents/{id}:
    get:
      summary: Consulta um documento
      security: [{ apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /documents/{id}/file:
    get:
      summary: Baixa o arquivo original
      security: [{ apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Bytes originais, Content-Type detectado pelo servidor.
          content:
            application/pdf: { schema: { type: string, format: binary } }
            image/jpeg: { schema: { type: string, format: binary } }
            image/png: { schema: { type: string, format: binary } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /documents/{id}/review:
    patch:
      summary: Aplica correção humana (concorrência otimista — ver ADR 0005)
      security: [{ apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version, reviewer]
              properties:
                version: { type: integer, description: "Version atual esperada do documento." }
                reviewer: { type: string }
                fields: { type: object, additionalProperties: { type: string } }
                docType: { $ref: "#/components/schemas/DocType" }
      responses:
        "200":
          description: Documento corrigido, status "reviewed".
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: >
            Dois casos distintos com o mesmo status HTTP, diferenciados por `error.code`:
            `version_conflict` (a versão enviada não bate com a atual — ADR 0005; corpo inclui
            `current` com o documento atual, para o cliente reconciliar) ou
            `invalid_review_state` (documento existe e a versão bate, mas `status` não é
            `needs_review`; corpo **não** inclui `current`, é erro de estado, não de concorrência).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties: { code: { type: string, enum: [version_conflict, invalid_review_state] }, message: { type: string } }
                  current:
                    allOf: [{ $ref: "#/components/schemas/Document" }]
                    description: "Presente só quando error.code = version_conflict."
  /health:
    servers:
      - url: http://localhost:3000
    get:
      summary: Health check — fora do prefixo /v1, sem API key (ADR 0006)
      security: []
      responses:
        "200":
          description: Processo está de pé.
          content:
            application/json:
              schema:
                type: object
                properties: { status: { type: string, enum: [ok] } }
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: Placeholder de autenticação serviço-a-serviço — ADR 0006.
  responses:
    BadRequest:
      description: Requisição inválida.
      content:
        application/json:
          schema:
            type: object
            properties: { error: { type: object } }
    Unauthorized:
      description: API key ausente ou inválida.
      content:
        application/json:
          schema:
            type: object
            properties: { error: { type: object } }
    NotFound:
      description: Documento não encontrado.
      content:
        application/json:
          schema:
            type: object
            properties: { error: { type: object } }
  schemas:
    Status:
      type: string
      enum: [received, processing, ready, needs_review, failed, reviewed]
    DocType:
      type: string
      enum: [identidade, comprovante_residencia, contracheque, outro]
    Document:
      type: object
      properties:
        id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/Status" }
        channel: { type: string, enum: [whatsapp, email, balcao, outro] }
        sourceFilename: { type: string }
        contentType: { type: string }
        sizeBytes: { type: integer }
        contentHash: { type: string }
        receivedAt: { type: string, format: date-time }
        docType: { $ref: "#/components/schemas/DocType" }
        fields: { type: object, additionalProperties: { type: string }, nullable: true }
        confidence: { type: number, minimum: 0, maximum: 1, nullable: true }
        suggestedFilename: { type: string, nullable: true }
        promptVersion: { type: string, nullable: true }
        modelVersion: { type: string, nullable: true }
        attempts: { type: integer }
        lastError: { type: string, nullable: true }
        reviewedBy: { type: string, nullable: true }
        reviewedAt: { type: string, format: date-time, nullable: true }
        version: { type: integer }
        duplicate: { type: boolean, description: "So presente na resposta de POST /documents." }